Animations for when moving player and boxes

+++ exam/rendering/ObjectRenderer.cpp
@@ -119,7 +119,7 @@ ObjectRenderer ObjectRenderer::create() {
 }
 
 void ObjectRenderer::draw(
-    glm::uvec2 position,
+    glm::vec2 position,
     glm::vec3 color,
     const framework::Texture *texture,
     const Light &light,
